requirements/dev.txt
-r base.txt

ruff==0.3.0

pre-commit==3.6.2

pytest==8.0.2
pytest-cov==4.1.0
pytest-django==4.8.0
pytest-sugar==1.0.0
